+++ packages/react-lowcode/src/codegen/facade/facadeApi.ts
@@ -14,7 +14,7 @@ import { CodeRW } from "../../io";
 import { CodegenRw } from "../io/codegenRw";
 import { BasicTableGenerator } from "../generation/generators/list/basic-table-generator";
 import { FacadeDeleteOptions, FacadeInsertOptions } from "./interfaces";
-import { ColumnSourcePositionResult } from "../interfaces";
+import { ColumnSourcePositionResult, WidgetProperties } from "../interfaces";
 
 export async function insertColumn(
   tablePosition: SourceLineCol,
@@ -196,4 +196,55 @@ export async function getColumnSourcePosition(
   );
   
   return await generator.getColumnSourcePosition(tablePosition, options.index!);
+}
+
+export async function getFormWidgetProperties(
+  position: SourceLineCol,
+  io: CodegenRw
+): Promise<WidgetProperties> {
+  let generationContext = {
+    uiFramework: UiFramework.MaterialUI,
+    formatter: Formatter.None,
+    index: { tableType: TableType.DataTable, height: "400px" },
+  };
+  let appContext = new AppContext(generationContext, io);
+  let sourceFileContext = new PageContext(
+    appContext,
+    position.fileName
+  );
+  let widgetContext = new WidgetContext(sourceFileContext);
+
+  let generator = new MuiDetailGenerator(
+    generationContext,
+    undefined,
+    widgetContext
+  );
+
+  return await generator.getFormWidgetProperties(position);
+}
+
+export async function setFormWidgetProperties(
+  position: SourceLineCol,
+  io: CodegenRw,
+  properties: WidgetProperties
+): Promise<string | undefined> {
+  let generationContext = {
+    uiFramework: UiFramework.MaterialUI,
+    formatter: Formatter.None,
+    index: { tableType: TableType.DataTable, height: "400px" },
+  };
+  let appContext = new AppContext(generationContext, io);
+  let sourceFileContext = new PageContext(
+    appContext,
+    position.fileName
+  );
+  let widgetContext = new WidgetContext(sourceFileContext);
+
+  let generator = new MuiDetailGenerator(
+    generationContext,
+    undefined,
+    widgetContext
+  );
+
+  return await generator.setFormWidgetProperties(position, properties);
 }